Rename "common" directory to "shared"
